wrap ioreport subscription and channels
